visitor: Fix build on 32 bit

+++ src/volt/visitor/prettyprinter.d
@@ -2198,7 +2198,7 @@ protected:
 		mSink(s);
 	}
 
-	void wf(size_t i)
+	void wf(ulong i)
 	{
 		string s = format("%s", i);
 		mSink(s);
